+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 # [8.6.0] - 2024-04-19
 - Added Experience Composer to Video API
-- Fixed regression in `createSession` endpoint which caused the API to return XML instead of JSON.
+- Fixed regression in `createSession` endpoint
+- Added Verify capability to Application API
 
 # [8.5.0] - 2024-04-12
 - Added Live Captions and Audio Connector endpoints to Video API